[Micro Update]
[File Actions.cs]
Added summaries to all the methods.
Changed in Download method the verification of the destination, as it's supposed to be a file, and to delete it if the file exists. This prevents a very important exception since windows doesn't damn clean up temporary files.

[First Start.cs]
Changed the text to warn the user that they should have enough space. Suggestion is 20GB, but yeah...

[Server Settings.cs, Workshop Download.cs, Server.cs]
Fixed exception due to missmatch in use of VerifyPath() methods.

[Servers.cs]
Same as above.
Will always attempt to update if it does not detect Unturned.exe in either VanillaServerPath or RocketmodServerPath.

[AssemblyInfo.cs]
Fixed copyright. Project was started and copyrighted in 2016, and still is copyrighted and receiving an update in 2018.

[SteamCMD.cs]
Fixed crash on RunCommand method due to it not running VerifySteam prior to any actions.
Added summaries to all methods.