Fixes & Workshop completed
[Workshop Download.cs]
Fully working edition.
Fixed crash when the Maps folder didn't exist.
Disables update/delete/view on steam buttons when empty.
Made DeleteAll_Click reload the entire list of installed items after deleting is complete.
Fixed View_Click so that it doesn't open a browser when no item is selected.
When loading all items, the first one will be automatically selected.
Fixed all Path.Combine as they used the root folder instead of the actual folder required.
When index is changed, it will disable the delete and view button accordingly.

[Server.cs]
Added summary to Folder property.

[SteamCMD.cs]
Fixed issue when copying the downloaded workshop folder, it would not copy due to missuse of Path.Combine and it would only copy the contents inside to the target directory, not to a directory with the exact same ID.